+++ src/main/scala/pro/civitaspo/digdag/plugin/ecs_task/EcsTaskPlugin.scala
@@ -8,7 +8,7 @@ import io.digdag.spi.{Operator, OperatorContext, OperatorFactory, OperatorProvid
 import javax.inject.Inject
 import pro.civitaspo.digdag.plugin.ecs_task.register.EcsTaskRegisterOperator
 import pro.civitaspo.digdag.plugin.ecs_task.result.EcsTaskResultOperator
-import pro.civitaspo.digdag.plugin.ecs_task.run.EcsTaskRunOperator
+import pro.civitaspo.digdag.plugin.ecs_task.run.{EcsTaskRunInternalOperator, EcsTaskRunOperator}
 import pro.civitaspo.digdag.plugin.ecs_task.wait.EcsTaskWaitOperator
 
 object EcsTaskPlugin {
@@ -23,6 +23,7 @@ object EcsTaskPlugin {
         operatorFactory("ecs_task.register", classOf[EcsTaskRegisterOperator]),
         operatorFactory("ecs_task.result", classOf[EcsTaskResultOperator]),
         operatorFactory("ecs_task.run", classOf[EcsTaskRunOperator]),
+        operatorFactory("ecs_task.run_internal", classOf[EcsTaskRunInternalOperator]),
         operatorFactory("ecs_task.wait", classOf[EcsTaskWaitOperator])
       )
     }
