+++ src/main/scala/pro/civitaspo/digdag/plugin/ecs_task/run/EcsTaskRunOperator.scala
@@ -1,213 +1,66 @@
 package pro.civitaspo.digdag.plugin.ecs_task.run
-import java.nio.file.Path
-
-import com.amazonaws.services.ecs.model.{
-  AwsVpcConfiguration,
-  ContainerOverride,
-  KeyValuePair,
-  NetworkConfiguration,
-  PlacementConstraint,
-  PlacementConstraintType,
-  PlacementStrategy,
-  PlacementStrategyType,
-  RunTaskRequest,
-  RunTaskResult,
-  TaskOverride
-}
 import com.google.common.base.Optional
-import com.google.common.collect.ImmutableList
-import io.digdag.client.config.{Config, ConfigException, ConfigKey}
-import io.digdag.spi.{ImmutableTaskResult, OperatorContext, PrivilegedVariables, SecretProvider, TaskRequest, TaskResult, TemplateEngine}
+import io.digdag.client.config.Config
+import io.digdag.spi.{OperatorContext, TaskResult, TemplateEngine}
 import io.digdag.util.DurationParam
 import pro.civitaspo.digdag.plugin.ecs_task.AbstractEcsTaskOperator
-import pro.civitaspo.digdag.plugin.ecs_task.register.EcsTaskRegisterOperator
-
-import scala.collection.JavaConverters._
 
 class EcsTaskRunOperator(operatorName: String, context: OperatorContext, systemConfig: Config, templateEngine: TemplateEngine)
     extends AbstractEcsTaskOperator(operatorName, context, systemConfig, templateEngine) {
 
-  // For the other operators configuration
-  val taskDefForRegister: Optional[Config] = params.getOptionalNested("def")
+  val cluster: String = params.get("cluster", classOf[String])
+  val taskDef: Optional[Config] = params.getOptionalNested("def")
   val resultS3UriPrefix: Optional[String] = params.getOptional("result_s3_uri_prefix", classOf[String])
   val timeout: DurationParam = params.get("timeout", classOf[DurationParam], DurationParam.parse("15m"))
 
-  // For ecs_task.run operator configuration
-  val cluster: String = params.get("cluster", classOf[String])
-  val count: Optional[Int] = params.getOptional("count", classOf[Int])
-  val group: Optional[String] = params.getOptional("group", classOf[String])
-  val launchType: Optional[String] = params.getOptional("launch_type", classOf[String])
-  val networkConfiguration: Optional[NetworkConfiguration] = configureNetworkConfiguration(params.getNestedOrGetEmpty("network_configuration"))
-  val overrides: Optional[TaskOverride] = configureTaskOverride(params.getNestedOrGetEmpty("overrides"))
-
-  val placementConstraints: Seq[PlacementConstraint] =
-    params.getListOrEmpty("placement_constraints", classOf[Config]).asScala.map(configurePlacementConstraint).map(_.get)
-
-  val placementStrategy: Seq[PlacementStrategy] =
-    params.getListOrEmpty("placement_strategy", classOf[Config]).asScala.map(configurePlacementStrategy).map(_.get)
-  val platformVersion: Optional[String] = params.getOptional("platform_version", classOf[String])
-  val startedBy: Optional[String] = params.getOptional("started_by", classOf[String])
-  val taskDefinition: Optional[String] = params.getOptional("task_definition", classOf[String]) // generated by ecs_task.register> operator if not set.
-
-  protected def buildRunTaskRequest(taskDefArn: String): RunTaskRequest = {
-    val req: RunTaskRequest = new RunTaskRequest()
-
-    req.setCluster(cluster)
-    if (count.isPresent) req.setCount(count.get)
-    if (group.isPresent) req.setGroup(group.get)
-    if (launchType.isPresent) req.setLaunchType(launchType.get)
-    if (networkConfiguration.isPresent) req.setNetworkConfiguration(networkConfiguration.get)
-    if (overrides.isPresent) req.setOverrides(overrides.get)
-    if (placementConstraints.nonEmpty) req.setPlacementConstraints(placementConstraints.asJava)
-    if (placementStrategy.nonEmpty) req.setPlacementStrategy(placementStrategy.asJava)
-    if (platformVersion.isPresent) req.setPlatformVersion(platformVersion.get)
-    if (startedBy.isPresent) req.setStartedBy(startedBy.get)
-    req.setTaskDefinition(taskDefArn)
-
-    req
-  }
-
-  protected def configureNetworkConfiguration(c: Config): Optional[NetworkConfiguration] = {
-    if (c.isEmpty) return Optional.absent()
-
-    val awsvpcConfiguration: Optional[AwsVpcConfiguration] = configureAwsVpcConfiguration(c.getNestedOrGetEmpty("awsvpc_configuration"))
-
-    val nc: NetworkConfiguration = new NetworkConfiguration()
-    if (awsvpcConfiguration.isPresent) nc.setAwsvpcConfiguration(awsvpcConfiguration.get)
-
-    Optional.of(nc)
-  }
-
-  protected def configureAwsVpcConfiguration(c: Config): Optional[AwsVpcConfiguration] = {
-    if (c.isEmpty) return Optional.absent()
-
-    val assignPublicIp: Optional[String] = c.getOptional("assign_public_ip", classOf[String])
-    val securityGroups: Seq[String] = c.getListOrEmpty("security_groups", classOf[String]).asScala
-    val subnets: Seq[String] = c.getListOrEmpty("subnets", classOf[String]).asScala
-
-    val avc: AwsVpcConfiguration = new AwsVpcConfiguration()
-    if (assignPublicIp.isPresent) avc.setAssignPublicIp(assignPublicIp.get)
-    if (securityGroups.nonEmpty) avc.setSecurityGroups(securityGroups.asJava)
-    if (subnets.nonEmpty) avc.setSubnets(subnets.asJava)
-
-    Optional.of(avc)
-  }
-
-  protected def configureTaskOverride(c: Config): Optional[TaskOverride] = {
-    if (c.isEmpty) return Optional.absent()
-
-    val containerOverrides: Seq[ContainerOverride] =
-      c.getListOrEmpty("container_overrides", classOf[Config]).asScala.map(configureContainerOverride).map(_.get)
-    val executionRoleArn: Optional[String] = c.getOptional("execution_role_arn", classOf[String])
-    val taskRoleArn: Optional[String] = c.getOptional("task_role_arn", classOf[String])
-
-    val to: TaskOverride = new TaskOverride()
-    if (containerOverrides.nonEmpty) to.setContainerOverrides(containerOverrides.asJava)
-    if (executionRoleArn.isPresent) to.setExecutionRoleArn(executionRoleArn.get)
-    if (taskRoleArn.isPresent) to.setTaskRoleArn(taskRoleArn.get)
-
-    Optional.of(to)
-  }
-
-  protected def configureContainerOverride(c: Config): Optional[ContainerOverride] = {
-    if (c.isEmpty) return Optional.absent()
-
-    val command: Seq[String] = c.getListOrEmpty("command", classOf[String]).asScala
-    val cpu: Optional[Int] = c.getOptional("cpu", classOf[Int])
-    val environment: Seq[KeyValuePair] = c
-      .getMapOrEmpty("environment", classOf[String], classOf[String])
-      .asScala
-      .map { case (k: String, v: String) => new KeyValuePair().withName(k).withValue(v) }
-      .toSeq // TODO: doc
-    val memory: Optional[Int] = c.getOptional("memory", classOf[Int])
-    val memoryReservation: Optional[Int] = c.getOptional("memory_reservation", classOf[Int])
-    val name: Optional[String] = c.getOptional("name", classOf[String])
-
-    val co: ContainerOverride = new ContainerOverride()
-    if (command.nonEmpty) co.setCommand(command.asJava)
-    if (cpu.isPresent) co.setCpu(cpu.get)
-    if (environment.nonEmpty) co.setEnvironment(environment.asJava)
-    if (memory.isPresent) co.setMemory(memory.get)
-    if (memoryReservation.isPresent) co.setMemoryReservation(memoryReservation.get)
-    if (name.isPresent) co.setName(name.get)
-
-    Optional.of(co)
-  }
-
-  protected def configurePlacementConstraint(c: Config): Optional[PlacementConstraint] = {
-    if (c.isEmpty) return Optional.absent()
-
-    val expression: Optional[String] = c.getOptional("expression", classOf[String])
-    val `type`: Optional[PlacementConstraintType] = c.getOptional("type", classOf[PlacementConstraintType])
-
-    val pc: PlacementConstraint = new PlacementConstraint()
-    if (expression.isPresent) pc.setExpression(expression.get)
-    if (`type`.isPresent) pc.setType(`type`.get)
-
-    Optional.of(pc)
-  }
-
-  protected def configurePlacementStrategy(c: Config): Optional[PlacementStrategy] = {
-    if (c.isEmpty) return Optional.absent()
-
-    val field: Optional[String] = c.getOptional("field", classOf[String])
-    val `type`: Optional[PlacementStrategyType] = c.getOptional("type", classOf[PlacementStrategyType])
-
-    val ps: PlacementStrategy = new PlacementStrategy()
-    if (field.isPresent) ps.setField(field.get)
-    if (`type`.isPresent) ps.setType(`type`.get)
-
-    Optional.of(ps)
-  }
-
-
   override def runTask(): TaskResult = {
-    if (!taskDefForRegister.isPresent && !taskDefinition.isPresent) throw new ConfigException("'def' or 'task_definition' must be set.")
-    if (taskDefForRegister.isPresent && taskDefinition.isPresent) throw new ConfigException("'def' and 'task_definition' cannot be used at the same time.")
+    val subTasks: Config = cf.create()
+    if (taskDef.isPresent) subTasks.setNested("+register", ecsTaskRegisterSubTask())
+    subTasks.setNested("+run", ecsTaskRunInternalSubTask())
+    subTasks.setNested("+wait", ecsTaskWaitSubTask())
 
-    val taskDefArn: String =
-      if (taskDefinition.isPresent) taskDefinition.get
-      else {
-        val taskResult: TaskResult = runEcsTaskRegister(taskDefForRegister.get)
-        taskResult.getStoreParams.getNested("last_ecs_task_register").get("task_definition_arn", classOf[String])
-      }
-
-    val req: RunTaskRequest = buildRunTaskRequest(taskDefArn)
-    logger.debug(req.toString)
-    val result: RunTaskResult = aws.withEcs(_.runTask(req))
-    logger.debug(result.toString)
-
-    val paramsToStore = cf.create()
-    val last_ecs_task_run: Config = paramsToStore.getNestedOrSetEmpty("last_ecs_task_run")
-    last_ecs_task_run.set("task_arns", result.getTasks.asScala.map(_.getTaskArn).asJava)
-
-    val builder: ImmutableTaskResult.Builder = TaskResult.defaultBuilder(cf)
-    builder.resetStoreParams(ImmutableList.of(ConfigKey.of("last_ecs_task_run")))
-    builder.storeParams(paramsToStore)
+    val builder = TaskResult.defaultBuilder(cf)
+    builder.subtaskConfig(subTasks)
     builder.build()
   }
 
-  protected def runEcsTaskRegister(registerConfig: Config): TaskResult = {
-    val config = params.deepCopy()
-    config.getKeys.asScala.filter(_.startsWith("_")).foreach(config.remove)
-    config.setNested("_command", registerConfig)
-    config.set("_type", "ecs_task.register")
-
-    val newContext: OperatorContext = new OperatorContext {
-      override def getProjectPath: Path = context.getProjectPath
-      override def getTaskRequest: TaskRequest =
-        TaskRequest
-          .builder()
-          .from(context.getTaskRequest)
-          .config(config)
-          .build()
-      override def getSecrets: SecretProvider = context.getSecrets
-      override def getPrivilegedVariables: PrivilegedVariables = context.getPrivilegedVariables
+  protected def ecsTaskRegisterSubTask(): Config = {
+    withDefaultSubTask { subTask =>
+      subTask.set("_type", "ecs_task.register")
+      subTask.set("_command", taskDef)
     }
+  }
 
-    val operator: EcsTaskRegisterOperator =
-      new EcsTaskRegisterOperator(operatorName = "ecs_task.register", context = newContext, systemConfig = systemConfig, templateEngine = templateEngine)
+  protected def ecsTaskRunInternalSubTask(): Config = {
+    val config: Config = params.deepCopy()
+    Seq("def", "result_s3_uri_prefix", "timeout").foreach(config.remove)
+    if (taskDef.isPresent) config.set("task_definition", "${last_ecs_task_register.task_definition_arn}")
+    withDefaultSubTask { subTask =>
+      subTask.set("_type", "ecs_task.run_internal")
+      subTask.set("_export", config)
+    }
+  }
 
-    operator.run()
+  protected def ecsTaskWaitSubTask(): Config = {
+    withDefaultSubTask { subTask =>
+      subTask.set("_type", "ecs_task.wait")
+      subTask.set("cluster", cluster)
+      subTask.set("tasks", "${last_ecs_task_run.task_arns}")
+      subTask.set("timeout", timeout.toString)
+    }
+  }
+
+  protected def withDefaultSubTask(f: Config => Config): Config = {
+    val subTask: Config = cf.create()
+
+    subTask.set("auth_method", aws.conf.authMethod)
+    subTask.set("profile_name", aws.conf.profileName)
+    if (aws.conf.profileFile.isPresent) subTask.set("profile_file", aws.conf.profileFile.get())
+    subTask.set("use_http_proxy", aws.conf.useHttpProxy)
+    if (aws.conf.region.isPresent) subTask.set("region", aws.conf.region.get())
+    if (aws.conf.endpoint.isPresent) subTask.set("endpoint", aws.conf.endpoint.get())
+
+    f(subTask)
+    subTask
   }
 }
